lexbridge/requirements.txt
# LexBridge — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core Framework ────────────────────────────────────────────────────────────
Flask==3.0.3
flask-cors==4.0.1

# ── Database ──────────────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.4.0
SQLAlchemy==2.0.31

# ── Authentication ────────────────────────────────────────────────────────────
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3          # password hashing (ships with Flask)

# ── Email ─────────────────────────────────────────────────────────────────────
Flask-Mail==0.10.0

# ── Rate Limiting ─────────────────────────────────────────────────────────────
Flask-Limiter==3.7.0
limits==3.13.0

# ── Environment Variables ─────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Production Server ─────────────────────────────────────────────────────────
gunicorn==22.0.0

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-flask==1.3.0
